backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai>=0.8.0
chromadb>=0.4.18
langchain==0.0.350
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pypdf2==3.0.1
python-docx==1.1.0
markdown==3.5.1
tiktoken>=0.5.2
numpy==1.24.3
pandas==2.1.3

# Free local models for testing (no API costs)
sentence-transformers==2.2.2
transformers==4.35.0
torch==2.1.0
